[update] ディレクトリ構成とファイル分けと名前変更
